empackage/requirements.txt
Fabric==1.9.1
PyYAML==3.11
click==3.2
ecdsa==0.11
fabtools==0.19.0
paramiko==1.14.1
pycrypto==2.6.1
wsgiref==0.1.2
